feat(marketplace-api): rename metadata endpoints
