+++ ark-marketplace-api/src/handlers/token_handler.rs
@@ -3,10 +3,9 @@ use super::utils::CHAIN_ID;
 use crate::db::db_access::DatabaseAccess;
 use crate::db::query::{
     flush_all_data_query, get_collection_floor_price, get_token_activity_data, get_token_data,
-    get_token_marketdata, get_token_offers_data, get_tokens_data, get_tokens_data_by_id,
-    get_tokens_portfolio_data, refresh_token_metadata,
+    get_token_marketdata, get_token_offers_data, get_tokens_data, get_tokens_portfolio_data,
+    refresh_token_metadata,
 };
-use crate::managers::elasticsearch_manager::ElasticsearchManager;
 use crate::models::token::TokenOfferOneData;
 use crate::models::token::{TokenEventType, TokenInformationData};
 use crate::utils::currency_utils::compute_floor_difference;
@@ -17,11 +16,8 @@ use redis::aio::MultiplexedConnection;
 use serde::Deserialize;
 use serde_json::json;
 use serde_qs;
-use serde_urlencoded;
-use std::collections::HashMap;
 use std::sync::Arc;
 use tokio::sync::Mutex;
-use urlencoding::decode;
 
 #[derive(Deserialize)]
 pub struct QueryParameters {
@@ -345,107 +341,6 @@ pub async fn post_refresh_token_metadata<D: DatabaseAccess + Sync>(
                 }
             }
         }
-        Err(err) => {
-            tracing::error!("error query get_tokens_data: {}", err);
-            HttpResponse::InternalServerError().finish()
-        }
-    }
-}
-
-pub async fn get_token_trait_filters<D: DatabaseAccess + Sync>(
-    path: web::Path<String>,
-    es_data: web::Data<HashMap<String, String>>,
-) -> impl Responder {
-    let contract_address = path.into_inner();
-    let elasticsearch_manager = ElasticsearchManager::new(es_data.get_ref().clone());
-
-    let normalized_address = normalize_address(&contract_address);
-    let result = elasticsearch_manager
-        .get_attributes_for_collection(&normalized_address, CHAIN_ID)
-        .await;
-
-    match result {
-        Ok(json_response) => HttpResponse::Ok().json(json!({
-            "data": json_response
-        })),
-        Err(e) => HttpResponse::InternalServerError().json(json!({
-            "error": format!("Failed to retrieve data: {}", e)
-        })),
-    }
-}
-
-pub async fn get_filtered_tokens<D: DatabaseAccess + Sync>(
-    req: HttpRequest,
-    path: web::Path<String>,
-    db_pool: web::Data<D>,
-    es_data: web::Data<HashMap<String, String>>,
-) -> impl Responder {
-    let query_string = req.query_string();
-    let query_params: HashMap<String, String> =
-        serde_urlencoded::from_str(query_string).unwrap_or_default();
-    let contract_address = path.into_inner();
-    let normalized_address = normalize_address(&contract_address);
-    let buy_now = query_params.get("buy_now").map(String::as_str) == Some("true");
-    let sort = query_params
-        .get("sort")
-        .map(String::as_str)
-        .unwrap_or("price");
-    let direction = query_params
-        .get("direction")
-        .map(String::as_str)
-        .unwrap_or("asc");
-
-    let (page, items_per_page) = match extract_page_params(query_string, 1, 100) {
-        Err(msg) => return HttpResponse::BadRequest().json(msg),
-        Ok((page, items_per_page)) => (page, items_per_page),
-    };
-
-    let mut token_ids = None;
-    if let Some(traits_param) = query_params.get("traits") {
-        let decoded_traits = decode(traits_param).expect("Failed to decode traits");
-        let traits_map: HashMap<String, Vec<String>> =
-            serde_json::from_str(&decoded_traits).expect("Failed to parse JSON");
-
-        let elasticsearch_manager = ElasticsearchManager::new(es_data.get_ref().clone());
-
-        let result = elasticsearch_manager
-            .search_tokens_by_traits(&normalized_address, CHAIN_ID, traits_map)
-            .await;
-
-        token_ids = match result {
-            Ok(token_ids) => Some(token_ids),
-            Err(e) => {
-                return HttpResponse::InternalServerError().json(json!({
-                    "error": format!("Failed to retrieve data: {}", e)
-                }))
-            }
-        };
-    }
-
-    let db_access = db_pool.get_ref();
-
-    match get_tokens_data_by_id(
-        db_access,
-        &normalized_address,
-        CHAIN_ID,
-        page,
-        items_per_page,
-        buy_now,
-        sort,
-        direction,
-        token_ids,
-    )
-    .await
-    {
-        Err(sqlx::Error::RowNotFound) => HttpResponse::NotFound().body("data not found"),
-        Ok((ref collection_data, _, _)) if collection_data.is_empty() => {
-            HttpResponse::NotFound().body("data not found")
-        }
-        Ok((collection_data, _has_next_page, token_count)) => HttpResponse::Ok().json(json!({
-            "data": collection_data,
-            "token_count": token_count,
-            "next_page": page + 1
-        })),
         Err(err) => {
             tracing::error!("error query get_tokens_data: {}", err);
             HttpResponse::InternalServerError().finish()
